+++ packages/rum/test/utils.ts
@@ -160,7 +160,7 @@ export function findFullSnapshot(segment: Segment): FullSnapshotRecord | null {
 }
 
 // Returns all the VisualViewportRecords in a Segment, if any.
-export function findAllVisualViewport(segment: Segment): VisualViewportRecord[] {
+export function findAllVisualViewports(segment: Segment): VisualViewportRecord[] {
   return segment.records.filter((record) => record.type === RecordType.VisualViewport) as VisualViewportRecord[]
 }
